refactor: move IdGenerator from domain to common
