+++ application/src/main/kotlin/com/dobby/usecase/experiment/CreateExperimentPostUseCase.kt
@@ -1,7 +1,7 @@
 package com.dobby.usecase.experiment
 
 import com.dobby.usecase.UseCase
-import com.dobby.IdGenerator
+import com.dobby.util.IdGenerator
 import com.dobby.gateway.experiment.ExperimentPostGateway
 import com.dobby.gateway.member.MemberGateway
 import com.dobby.model.experiment.ApplyMethod
