+++ application/src/main/kotlin/com/dobby/usecase/member/email/SendEmailCodeUseCase.kt
@@ -2,7 +2,7 @@ package com.dobby.usecase.member.email
 
 import com.dobby.usecase.UseCase
 import com.dobby.EmailTemplateLoader
-import com.dobby.IdGenerator
+import com.dobby.util.IdGenerator
 import com.dobby.gateway.CacheGateway
 import com.dobby.gateway.email.EmailGateway
 import com.dobby.gateway.email.VerificationGateway
