+++ common/src/main/kotlin/com/dobby/util/IdGenerator.kt
@@ -1,4 +1,4 @@
-package com.dobby
+package com.dobby.util
 
 interface IdGenerator {
     fun generateId(): String
