+++ infrastructure/src/main/kotlin/com/dobby/external/s3/S3PreSignedUrlProvider.kt
@@ -1,7 +1,7 @@
 package com.dobby.external.s3
 
 import com.dobby.exception.InvalidRequestValueException
-import com.dobby.IdGenerator
+import com.dobby.util.IdGenerator
 import com.dobby.config.properties.S3Properties
 import org.springframework.stereotype.Component
 import software.amazon.awssdk.services.s3.presigner.S3Presigner
