+++ infrastructure/src/main/kotlin/com/dobby/persistence/identifier/TsidIdGenerator.kt
@@ -1,6 +1,6 @@
 package com.dobby.persistence.identifier
 
-import com.dobby.IdGenerator
+import com.dobby.util.IdGenerator
 import com.github.f4b6a3.tsid.TsidCreator
 import org.springframework.stereotype.Component
