+++ infrastructure/src/test/kotlin/com/dobby/external/s3/S3PreSignedUrlProviderTest.kt
@@ -1,7 +1,7 @@
 package com.dobby.external.s3
 
 import com.dobby.exception.InvalidRequestValueException
-import com.dobby.IdGenerator
+import com.dobby.util.IdGenerator
 import com.dobby.config.properties.S3Properties
 import io.kotest.core.spec.style.BehaviorSpec
 import io.mockk.every
